requirements.txt
aenum==3.1.15
ai-api-client-sdk==1.28.0
aiofiles==23.1.0
aiohttp==3.8.5
aiosignal==1.3.1
altair==5.0.1
anyio==3.7.1
argilla==0.0.1
asgiref==3.6.0
astroid==2.15.6
async-timeout==4.0.2
attrs==23.1.0
awscli==1.22.101
backoff==2.2.1
blis==0.7.10
boto3==1.21.46
botocore==1.24.46
cachetools==5.3.1
catalogue==2.0.9
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
chromadb==0.3.23
click==8.0.4
clickhouse-connect==0.6.8
colorama==0.4.3
confection==0.1.1
contourpy==1.0.7
cryptography==41.0.3
cycler==0.11.0
cymem==2.0.7
dataclasses-json==0.5.14
dill==0.3.7
Django==4.2.1
docker==4.3.1
dockerpty==0.4.1
docopt==0.6.2
docutils==0.15.2
duckdb==0.8.1
et-xmlfile==1.1.0
exceptiongroup==1.1.2
fastapi==0.101.0
ffmpy==0.3.1
filelock==3.12.2
fonttools==4.39.4
frozenlist==1.4.0
fsspec==2023.6.0
gitdb==4.0.10
GitPython==3.1.7
google-auth==2.22.0
gradio_client==0.3.0
h11==0.14.0
hnswlib==0.7.0
httpcore==0.17.3
httptools==0.6.0
httpx==0.24.1
huggingface-hub==0.16.4
idna==3.4
importlib-metadata==6.8.0
importlib-resources==5.12.0
isort==5.12.0
japanize-matplotlib==1.1.3
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
jsonschema==4.18.6
jsonschema-specifications==2023.7.1
kiwisolver==1.4.4
kubernetes==27.2.0
langchain==0.0.172
langcodes==3.3.0
langsmith==0.0.19
lazy-object-proxy==1.9.0
libmagic==1.0
linkify-it-py==2.0.2
lxml==4.9.3
lz4==4.3.2
Markdown==3.4.4
markdown-it-py==2.2.0
MarkupSafe==2.1.3
marshmallow==3.20.1
matplotlib==3.7.1
mccabe==0.7.0
mdit-py-plugins==0.3.3
mdurl==0.1.2
metaflow==2.8.0
monotonic==1.6
mpmath==1.3.0
msg-parser==1.2.0
multidict==6.0.4
murmurhash==1.0.9
mypy-extensions==1.0.0
networkx==3.1
nltk==3.8.1
numexpr==2.8.4
numpy==1.24.3
oauthlib==3.2.2
olefile==0.46
openai==0.27.6
openapi-schema-pydantic==1.2.4
openpyxl==3.1.2
orjson==3.9.2
packaging==23.1
pandas==2.0.1
pathy==0.10.2
pdf2image==1.16.3
pdfminer.six==20221105
Pillow==9.5.0
platformdirs==3.10.0
posthog==3.0.1
preshed==3.0.8
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==1.10.12
pydub==0.25.1
Pygments==2.15.1
pyhumps==3.8.0
pykwalify==1.7.0
pylint==2.17.5
pypandoc==1.11
pyparsing==3.0.9
pytesseract==0.3.10
python-box==5.1.1
python-dateutil==2.8.2
python-docx==0.8.11
python-dotenv==0.19.2
python-magic==0.4.27
python-multipart==0.0.6
python-pptx==0.6.21
pytz==2023.3
PyYAML==5.4.1
referencing==0.30.2
regex==2023.6.3
requests==2.31.0
requests-oauthlib==1.3.1
rpds-py==0.9.2
rsa==4.7.2
s3transfer==0.5.2
safetensors==0.3.1
sap-ai-core-metaflow==1.1.14
scikit-learn==1.2.2
scipy==1.10.1
seaborn==0.12.2
semantic-version==2.10.0
semver==3.0.1
sentence-transformers==2.2.2
sentencepiece==0.1.99
six==1.16.0
smart-open==6.3.0
smmap==5.0.0
sniffio==1.3.0
spacy==3.6.0
spacy-legacy==3.0.12
spacy-loggers==1.0.4
spython==0.0.79
SQLAlchemy==2.0.19
sqlparse==0.4.4
srsly==2.4.7
starlette==0.27.0
sympy==1.12
tabulate==0.9.0
tenacity==8.2.2
thinc==8.1.10
threadpoolctl==3.1.0
tiktoken==0.4.0
tokenizers==0.13.3
tomli==2.0.1
tomlkit==0.12.1
toolz==0.12.0
torch==2.0.1
torchvision==0.15.2
tqdm==4.65.0
transformers==4.31.0
typer==0.9.0
typing-inspect==0.9.0
typing_extensions==4.7.1
tzdata==2023.3
uc-micro-py==1.0.2
unstructured==0.6.6
urllib3==1.26.16
uvicorn==0.23.2
uvloop==0.17.0
wasabi==1.1.2
watchfiles==0.19.0
websocket-client==1.6.1
websockets==11.0.3
wrapt==1.15.0
XlsxWriter==3.1.2
yarl==1.9.2
zipp==3.15.0
zstandard==0.21.0
